+++ console/src/project/RepositoryRefreshPopup/actions.js
@@ -18,22 +18,25 @@
  * =====
  */
 // @flow
+import type {ConcordKey} from "../../types";
 
 const NAMESPACE = "repository";
 
 const types = {
-    REPOSITORY_TEST_REQUEST: `${NAMESPACE}/test/request`,
-    REPOSITORY_TEST_RESPONSE: `${NAMESPACE}/test/response`,
-    REPOSITORY_TEST_RESET: `${NAMESPACE}/test/reset`
+    REPOSITORY_REFRESH_REQUEST: `${NAMESPACE}/refresh/request`,
+    REPOSITORY_REFRESH_RESPONSE: `${NAMESPACE}/refresh/response`,
+    REPOSITORY_REFRESH_RESET: `${NAMESPACE}/refresh/reset`
 };
 
 export default types;
 
-export const testRepository = (data: any) => ({
-    type: types.REPOSITORY_TEST_REQUEST,
-    data
+export const refreshRepository = (orgName: ConcordKey, projectName: ConcordKey, repositoryName: ConcordKey) => ({
+    type: types.REPOSITORY_REFRESH_REQUEST,
+    orgName,
+    projectName,
+    repositoryName
 });
 
-export const resetTest = () => ({
-    type: types.REPOSITORY_TEST_RESET
+export const resetRefresh = () => ({
+    type: types.REPOSITORY_REFRESH_RESET
 });