+++ console/src/project/RepositoryRefreshPopup/reducers.js
@@ -21,49 +21,50 @@
 import {combineReducers} from "redux";
 import types from "./actions";
 
-const testResult = (state: any = null, {type, response, error}: any) => {
+const refreshResult = (state: any = null, {type, response, error}: any) => {
     switch (type) {
-        case types.REPOSITORY_TEST_RESPONSE:
+        case types.REPOSITORY_REFRESH_RESPONSE:
             if (error) {
                 return false;
             }
             return response;
-        case types.REPOSITORY_TEST_REQUEST:
-        case types.REPOSITORY_TEST_RESET:
+        case types.REPOSITORY_REFRESH_REQUEST:
+        case types.REPOSITORY_REFRESH_RESET:
             return null;
         default:
             return state;
     }
 };
 
-const loading = (state = false, {type}: any) => {
+const refreshing = (state = false, {type}: any) => {
     switch (type) {
-        case types.REPOSITORY_TEST_REQUEST:
+        case types.REPOSITORY_REFRESH_REQUEST:
             return true;
-        case types.REPOSITORY_TEST_RESPONSE:
-        case types.REPOSITORY_TEST_RESET:
+        case types.REPOSITORY_REFRESH_RESPONSE:
+        case types.REPOSITORY_REFRESH_RESET:
             return false;
         default:
             return state;
     }
 };
 
-const error = (state: any = null, {type, error, message}: any) => {
+const refreshError = (state: any = null, {type, error, message}: any) => {
     switch (type) {
-        case types.REPOSITORY_TEST_RESPONSE:
+        case types.REPOSITORY_REFRESH_RESPONSE:
             if (!error) {
                 return null;
             }
             return message;
-        case types.REPOSITORY_TEST_REQUEST:
-        case types.REPOSITORY_TEST_RESET:
+        case types.REPOSITORY_REFRESH_REQUEST:
+        case types.REPOSITORY_REFRESH_RESET:
             return null;
         default:
             return state;
     }
 };
-export default combineReducers({testResult, loading, error});
 
-export const getTestResult = (state: any) => state.testResult;
-export const isLoading = (state: any) => state.loading;
-export const getError = (state: any) => state.error;
+export default combineReducers({refreshResult, refreshing, refreshError});
+
+export const isRefreshing = (state: any) => state && state.refreshing;
+export const getRefreshResult = (state: any) => state && state.refreshResult;
+export const getRefreshError = (state: any) => state && state.refreshError;